# apps/web/openapi.json
{
  "components": {
    "schemas": {
      "AgentItem": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instruction_version": {
            "title": "Instruction Version",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "role",
          "provider",
          "model",
          "status",
          "instruction_version"
        ],
        "title": "AgentItem",
        "type": "object"
      },
...
      "DatabaseProvisioningPolicy": {
        "additionalProperties": false,
        "description": "Opt-in database provisioning without storing an administrator secret.",
        "properties": {
          "admin_url_secret_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Url Secret Reference"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "injected_environment_key": {
            "default": "DATABASE_URL",
            "title": "Injected Environment Key",
            "type": "string"
          }
        },
        "title": "DatabaseProvisioningPolicy",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListPage_AgentItem_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "offset",
          "limit",
          "truncated"
        ],
        "title": "ListPage[AgentItem]",
        "type": "object"
      },
      "ListPage_ApprovalItem_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApprovalItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "offset",
          "limit",
          "truncated"
        ],
        "title": "ListPage[ApprovalItem]",
        "type": "object"
      },
      "ListPage_AuditItem_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "offset",
          "limit",
          "truncated"
        ],
        "title": "ListPage[AuditItem]",
        "type": "object"
      },
      "ListPage_EvaluationItem_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvaluationItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "offset",
          "limit",
          "truncated"
        ],
        "title": "ListPage[EvaluationItem]",
        "type": "object"
      },
      "ListPage_PermissionItem_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PermissionItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "offset",
          "limit",
          "truncated"
        ],
        "title": "ListPage[PermissionItem]",
        "type": "object"
      },
      "ListPage_UsageItem_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UsageItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "offset",
          "limit",
          "truncated"
        ],
        "title": "ListPage[UsageItem]",
        "type": "object"
      },
...
      "PolicyProjection": {
        "additionalProperties": false,
        "properties": {
          "database_enabled": {
            "title": "Database Enabled",
            "type": "boolean"
          },
          "limits": {
            "$ref": "#/components/schemas/PolicyLimits"
          },
          "policy_digest": {
            "title": "Policy Digest",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "runner_mode": {
            "title": "Runner Mode",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "version",
          "policy_digest",
          "runner_mode",
          "database_enabled",
          "limits"
        ],
        "title": "PolicyProjection",
        "type": "object"
      },
      "ProjectCreateRequest": {
        "additionalProperties": false,
        "description": "Closed project registration body.",
        "properties": {
          "allowed_environment_files": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Allowed Environment Files",
            "type": "array"
          },
          "allowed_merge_methods": {
            "default": [
              "squash"
            ],
            "items": {
              "type": "string"
            },
            "title": "Allowed Merge Methods",
            "type": "array"
          },
          "commands": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CommandSpec"
            },
            "title": "Commands",
            "type": "array"
          },
          "database": {
            "$ref": "#/components/schemas/DatabaseProvisioningPolicy"
          },
          "default_branch": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Default Branch",
            "type": "string"
          },
          "developer_model": {
            "$ref": "#/components/schemas/AgentModelPolicy"
          },
          "github_repository": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Github Repository",
            "type": "string"
          },
          "instructions_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions Path"
          },
          "local_remediation_limit": {
            "default": 3,
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Local Remediation Limit",
            "type": "integer"
          },
          "merge_blocking_severities": {
            "default": [
              "blocker",
              "major"
            ],
            "items": {
              "$ref": "#/components/schemas/FindingSeverity"
            },
            "title": "Merge Blocking Severities",
            "type": "array",
            "uniqueItems": true
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "planner_model": {
            "$ref": "#/components/schemas/AgentModelPolicy"
          },
          "publication_blocking_severities": {
            "default": [
              "blocker",
              "major"
            ],
            "items": {
              "$ref": "#/components/schemas/FindingSeverity"
            },
            "title": "Publication Blocking Severities",
            "type": "array",
            "uniqueItems": true
          },
          "remote_remediation_limit": {
            "default": 3,
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Remote Remediation Limit",
            "type": "integer"
          },
          "repository_path": {
            "minLength": 1,
            "title": "Repository Path",
            "type": "string"
          },
          "reviewer_model": {
            "$ref": "#/components/schemas/AgentModelPolicy"
          },
          "runner_mode": {
            "$ref": "#/components/schemas/RunnerMode",
            "default": "docker"
          },
          "secret_paths": {
            "default": [
              ".env",
              ".env.local"
            ],
            "items": {
              "type": "string"
            },
            "title": "Secret Paths",
            "type": "array"
          },
          "trusted_project": {
            "default": false,
            "title": "Trusted Project",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "repository_path",
          "github_repository",
          "default_branch"
        ],
        "title": "ProjectCreateRequest",
        "type": "object"
      },
...
      "ProjectResponse": {
        "additionalProperties": false,
        "description": "Safe project identity and current policy summary.",
        "properties": {
          "canonical_path_key": {
            "title": "Canonical Path Key",
            "type": "string"
          },
          "default_branch": {
            "title": "Default Branch",
            "type": "string"
          },
          "github_repository": {
            "title": "Github Repository",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instructions_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions Path"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy"
          },
          "policy_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Digest"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "repository_path": {
            "title": "Repository Path",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "repository_path",
          "canonical_path_key",
          "github_repository",
          "default_branch",
          "instructions_path",
          "policy_version",
          "policy_digest"
        ],
        "title": "ProjectResponse",
        "type": "object"
      },
...
      "RunProjection": {
        "additionalProperties": false,
        "properties": {
          "agents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentSection"
            },
            "title": "Agents",
            "type": "object"
          },
          "available_commands": {
            "items": {
              "$ref": "#/components/schemas/AvailableCommand"
            },
            "title": "Available Commands",
            "type": "array"
          },
          "budgets": {
            "$ref": "#/components/schemas/BudgetSection"
          },
          "candidate": {
            "$ref": "#/components/schemas/CandidateSection"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckItem"
            },
            "title": "Checks",
            "type": "array"
          },
          "latest_events": {
            "items": {
              "$ref": "#/components/schemas/EventItem"
            },
            "title": "Latest Events",
            "type": "array"
          },
          "next_gate": {
            "anyOf": [
              {
                "enum": [
                  "plan",
                  "pr",
                  "merge"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Gate"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanSection"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectSection"
          },
          "pull_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PullRequestSection"
              },
              {
                "type": "null"
              }
            ]
          },
          "resource": {
            "$ref": "#/components/schemas/ResourceSection"
          },
          "review": {
            "$ref": "#/components/schemas/ReviewSection"
          },
          "run": {
            "$ref": "#/components/schemas/RunResponse"
          },
          "security": {
            "$ref": "#/components/schemas/SecuritySection"
          },
          "task": {
            "$ref": "#/components/schemas/TaskSection"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageSummary"
          }
        },
        "required": [
          "run",
          "task",
          "project",
          "resource",
          "plan",
          "candidate",
          "pull_request",
          "checks",
          "review",
          "agents",
          "budgets",
          "usage",
          "security",
          "latest_events",
          "available_commands",
          "next_gate"
        ],
        "title": "RunProjection",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "CSRF": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "OperatorSession": {
        "in": "cookie",
        "name": "forge_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Parallel Forge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agents": {
      "get": {
        "operationId": "agents_api_agents_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_AgentItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Agents"
      }
    },
    "/api/approvals": {
      "get": {
        "operationId": "approvals_api_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_ApprovalItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Approvals"
      }
    },
...
    "/api/audit": {
      "get": {
        "operationId": "audit_api_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_AuditItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Audit"
      }
    },
...
    "/api/evaluations": {
      "get": {
        "operationId": "evaluations_api_evaluations_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_EvaluationItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Evaluations"
      }
    },
...
    "/api/projects": {
      "get": {
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response List Projects Api Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "List Projects"
      },
      "post": {
        "operationId": "register_project_api_projects_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRF": [],
            "OperatorSession": []
          }
        ],
        "summary": "Register Project"
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "operationId": "get_project_api_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Get Project"
      }
    },
    "/api/projects/{project_id}/policy-projection": {
      "get": {
        "operationId": "policy_api_projects__project_id__policy_projection_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyProjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Policy"
      }
    },
...
    "/api/runs/{run_id}/projection": {
      "get": {
        "operationId": "projection_api_runs__run_id__projection_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunProjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Projection"
      }
    },
...
    "/api/tool-permissions": {
      "get": {
        "operationId": "permissions_api_tool_permissions_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_PermissionItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Permissions"
      }
    },
    "/api/usage": {
      "get": {
        "operationId": "usage_api_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_UsageItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Usage"
      }
    }
  }
}